# rest/src/main/resources/openapi/document-upload.yaml
openapi: 3.0.3
info:
  title: Paperless – Document API
  version: 1.0.0
servers:
  - url: http://localhost:8081
paths:
  /api/documents:
    post:
      summary: Upload a document
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: Resource URL of the created document
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Bad request
    get:
      summary: List recent documents
      operationId: listDocuments
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Maximum number of items to return
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
  /api/documents/{id}:
    get:
      summary: Get a document by id
      operationId: getDocumentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Not found
  /api/search:
    get:
      summary: Search documents
      operationId: searchDocuments
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
          description: Query string
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
components:
  schemas:
    Document:
      type: object
      properties:
        id:
          type: integer
          format: int64
        filename:
          type: string
        contentType:
          type: string
        size:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
      required: [id, filename, contentType, size, uploadedAt]